Combine both integration tests
