+++ test/dpul_collections/indexing_pipeline/integration/figgy_transformer_integration_test.exs
@@ -1,29 +1,8 @@
 defmodule DpulCollections.IndexingPipeline.FiggyTransformerIntegrationTest do
   use DpulCollections.DataCase
 
-  alias DpulCollections.{FiggyRepo, Repo}
-
-  alias DpulCollections.IndexingPipeline.{
-    FiggyHydrator,
-    FiggyResource,
-    FiggyTransformer,
-    HydrationCacheEntry,
-    TransformationCacheEntry
-  }
-
+  alias DpulCollections.IndexingPipeline.FiggyTransformer
   alias DpulCollections.IndexingPipeline
-
-  def start_figgy_producer(batch_size \\ 1) do
-    {:ok, hydrator} =
-      FiggyHydrator.start_link(
-        cache_version: 0,
-        producer_module: FiggyTestProducer,
-        producer_options: {self()},
-        batch_size: batch_size
-      )
-
-    hydrator
-  end
 
   def start_transformer_producer(batch_size \\ 1) do
     pid = self()
@@ -198,56 +177,5 @@ defmodule DpulCollections.IndexingPipeline.FiggyTransformerIntegrationTest do
     # Ensure there are no transformation cache entries.
     entries = IndexingPipeline.list_transformation_cache_entries()
     assert length(entries) == 0
-  end
-
-  def wait_for_hydrated_id(id, cache_version \\ 0) do
-    case IndexingPipeline.get_processor_marker!("hydrator", 0) do
-      %{cache_record_id: ^id} ->
-        true
-
-      _ ->
-        :timer.sleep(50)
-        wait_for_hydrated_id(id, cache_version)
-    end
-  end
-
-  def wait_for_transformed_id(id, cache_version \\ 0) do
-    case IndexingPipeline.get_processor_marker!("transformer", 0) do
-      %{cache_record_id: ^id} ->
-        true
-
-      _ ->
-        :timer.sleep(50)
-        wait_for_transformed_id(id, cache_version)
-    end
-  end
-
-  test "a full hydrator and transformer run" do
-    # Start the figgy producer
-    hydrator = start_figgy_producer(50)
-    # Demand all of them.
-    count = FiggyRepo.aggregate(FiggyResource, :count)
-    FiggyTestProducer.process(count)
-    # Wait for the last ID to show up.
-    task = Task.async(fn -> wait_for_hydrated_id(FiggyTestSupport.last_marker().id) end)
-    Task.await(task, 15000)
-    :timer.sleep(2000)
-    hydrator |> Broadway.stop(:normal)
-
-    # Start the transformer producer
-    transformer = start_transformer_producer(50)
-    entry_count = Repo.aggregate(HydrationCacheEntry, :count)
-    TestFiggyTransformerProducer.process(entry_count)
-    # Wait for the last ID to show up.
-    task =
-      Task.async(fn ->
-        wait_for_transformed_id(FiggyTestSupport.last_hydration_cache_entry_marker().id)
-      end)
-
-    Task.await(task, 15000)
-    transformation_cache_entry_count = Repo.aggregate(TransformationCacheEntry, :count)
-    assert FiggyTestSupport.ephemera_folder_count() == transformation_cache_entry_count
-    :timer.sleep(2000)
-    transformer |> Broadway.stop(:normal)
   end
 end
